# backend/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: pyq
    description: PYQ download operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /pyq/check:
    post:
      operationId: checkPyq
      tags: [pyq]
      summary: Check available PYQs
      description: Returns list of available PYQ URLs without downloading
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PyqInput"
      responses:
        "200":
          description: Available papers found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PyqCheckResult"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /pyq/download:
    post:
      operationId: downloadPyq
      tags: [pyq]
      summary: Download PYQs as ZIP
      description: Downloads all available PYQs and streams a ZIP file
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PyqInput"
      responses:
        "200":
          description: ZIP file containing all found PYQs
          content:
            application/zip:
              schema:
                type: string
                format: binary
        "404":
          description: No papers found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    PyqInput:
      type: object
      required:
        - subjectCode
        - course
        - startYear
        - endYear
        - sessions
      properties:
        subjectCode:
          type: string
          description: GTU subject code (e.g. 3150703)
        course:
          type: string
          description: Course name (e.g. BE, BCA, MCA, Diploma)
        startYear:
          type: integer
          description: Start year for range (e.g. 2021)
        endYear:
          type: integer
          description: End year for range (e.g. 2025)
        sessions:
          type: array
          items:
            type: string
          description: Session types - S for Summer, W for Winter

    PyqCheckResult:
      type: object
      required:
        - availableUrls
        - totalFound
      properties:
        availableUrls:
          type: array
          items:
            type: string
          description: List of available PDF URLs
        totalFound:
          type: integer
          description: Number of papers found
        checkedUrls:
          type: integer
          description: Total URLs checked

    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
        details:
          type: ["string", "null"]